+++ draft-js-plugins-editor/CHANGELOG.md
@@ -3,10 +3,16 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## To Be Released
+## 2.0.4
+- Tighten Immutable.js dependency requirements to ~3.7.4 to match draft-js and other plugins
 
-### Added
+## 2.0.3
+- Bugfix - componentWillReceiveProps causes infite update loop in some circumstances
 
+## 2.0.2
+- Automatically update decoratorless editorState upon props update
+
+## 2.0.1
 - blockStyleFn returns '' instead of false
 - `handleKeyCommand` now receives the arguments `(command, editorState, pluginFunctions)`
 - `handlePastedText` now receives the arguments `(text, html, editorState, pluginFunctions)`